% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod03_add_corine.R
\name{add_corine}
\alias{add_corine}
\title{Add CORINE}
\usage{
add_corine(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_03.RDS file to project folder
}
\description{
This function adds CORINE land cover data to the basemap.
}
