% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod06_add_CROME.R
\name{add_CROME}
\alias{add_CROME}
\title{Add Crop Map of England data}
\usage{
add_CROME(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_06.RDS file to project folder
}
\description{
This function adds Crop Map of England data to the basemap (England only).
}
