% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod07_add_DTM.R
\name{add_DTM}
\alias{add_DTM}
\title{Add terrain data}
\usage{
add_DTM(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_07.RDS file to project folder
}
\description{
This function adds terrain data (elevation) to the basemap. Data input can be one or several rasters of tif or asc format.
}
