% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod08_add_hedgerows.R
\name{add_hedgerows}
\alias{add_hedgerows}
\title{Add Hedgerows}
\usage{
add_hedgerows(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_08.RDS file to project folder
}
\description{
This function adds hedgerow data to the basemap. As opposed to the other basemap processing functions that extract information into existing polygons, this function creates new geometries within the basemap. Data input for hedgerows must be linear features.
}
