% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod04_add_NFI.R
\name{add_NFI}
\alias{add_NFI}
\title{Add National Forest Inventory}
\usage{
add_NFI(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_04.RDS file to project folder
}
\description{
This function adds National Forest Inventory data to the basemap (England).
}
