% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod05_add_PHI.R
\name{add_PHI}
\alias{add_PHI}
\title{Add Priority Habitat Inventory}
\usage{
add_PHI(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_05.RDS file to project folder
}
\description{
This function adds the Priority Habitat Inventory data to the basemap (England).
}
