% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod10_add_socioeco.R
\name{add_socioeco}
\alias{add_socioeco}
\title{Add socio-economic data}
\usage{
add_socioeco(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment. Needs to have undergone classification as population data are extracted into polygons classified as domestic buildings.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_classified_pop.RDS file to project folder
}
\description{
This function adds population (UK Census 2011) and health (Index of Multiple Deprivation 2019) data into the basemap where people are living (domestic buildings). The resulting fields are housePop (average number of people per dwelling )
}
