% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tiling_extract.R
\name{addAttributes}
\alias{addAttributes}
\title{Add Attributes}
\usage{
addAttributes(mm, colname, key)
}
\arguments{
\item{mm}{The mastermap object, a list of sf tiles}

\item{colname}{The name of the column that needs updating}

\item{key}{The raster attribute table or lookup table allowing to match the numeric code to the appropriate level.}
}
\value{
An updated mm object
}
\description{
Replace the information extracted in a raster by its text description
}
