% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{calculateDistances}
\alias{calculateDistances}
\title{Calculate Euclidean Distance}
\usage{
calculateDistances(r)
}
\arguments{
\item{r}{a raster with values of 1 (target pixels) and 8888 (pixels to calculate distances)}
}
\value{
a distance raster where the value is the distance (in m) to a source. Sources have a value of 0.
}
\description{
Calculate distance to a source (e.g. a habitat type, roads, an airport) when the landscape can be converted into binary data: source = 1, pixels to compute distance for = 8888 (and rest NA, masked by a buffer layer).
}
