% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_access_nature.R
\name{capacity_access_nature}
\alias{capacity_access_nature}
\title{Access to Nature Capacity Model}
\usage{
capacity_access_nature(
  x = parent.frame()$mm,
  studyArea = parent.frame()$studyArea,
  prow,
  crow,
  lnr,
  nnr,
  res = 5,
  local = 300,
  threshold = 500,
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attributes HabCode_B, Group (OS MasterMap Descriptive Group), GI (type of green infrastructure), and GIpublic (Private/Restricted/Public access).}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{prow}{Path to folder containing Public Rights of Way data for your study area (shapefile or geopackage).}

\item{crow}{Path to folder containing CRoW data (Open Country and Registered Common Land) for your study area (shapefile or geopackage).}

\item{lnr}{Path to folder containing Local Nature Reserves boundaries for your study area (shapefile or geopackage).}

\item{nnr}{Path to folder containing National Nature Reserves boundaries for your study area (shapefile or geopackage).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{local}{Radius (m) for focal statistics at local range (maximum distance for effectiveness). Default is 100 m.}

\item{threshold}{Size (m2) below which an isolated patch is not considered able to provide the service. Default is 100 m2.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".
#' @param save Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with capacity scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum capacity for the area).
}
\description{
Runs the access to nature ecosystem service model, generating capacity scores based on perceived "naturalness" of the habitats and their public accessibility. Calls external datasets which must be supplied by user: Public Rights of Way, CRoW, Local and National Nature Reserves.
}
