% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_climate_reg.R
\name{capacity_climate_reg}
\alias{capacity_climate_reg}
\title{Climate Regulation Capacity Model}
\usage{
capacity_climate_reg(
  x = parent.frame()$mm,
  studyArea = parent.frame()$studyArea,
  res = 5,
  local = 200,
  use_hedges = FALSE,
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attribute HabCode_B.}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{local}{Radius (m) for focal statistics at local range (maximum distance for effectiveness). Default is 300 m.}

\item{use_hedges}{Use a separate hedgerow layer? Default FALSE, see clean_hedgerows() for producing a model-ready hedge layer.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".}

\item{save}{Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with capacity scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum capacity for the area).
}
\description{
Runs the climate regulation ecosystem service model, generating capacity scores based on the ability of vegetation and water bodies to cool down the temperature locally.
}
