% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_pollination.R
\name{capacity_pollination}
\alias{capacity_pollination}
\title{Pollination Capacity Model}
\usage{
capacity_pollination(
  x,
  studyArea,
  res = 5,
  dist = 20,
  elev = "elev",
  DTM = NULL,
  elev_threshold = 250,
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attribute HabCode_B.}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{dist}{Distance within which pollinators are found in edge habitats (woodland) from a core habitat. Default 20 m.}

\item{elev}{A character string giving the name of the basemap attribute containing a polygon's mean elevation. The output raster will be masked to reflect the fact that pollinators are not abundant at high elevations. If the information has not been extracted in the basemap, a DTM can be supplied separately.}

\item{DTM}{A raster of elevation values for the study area (specified as file path to a folder). If left NULL (default), will use the basemap "elev" information instead (quicker).}

\item{elev_threshold}{The elevation (in m) threshold above which the model considers that pollinators are absent. Default 250 m.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".
#' @param save Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with capacity scores: one with raw scores (0-1: likelihood that a pollinator will visit a pixel), and one rescaled 0-100 (where 100 is maximum capacity for the area).
}
\description{
Runs the pollination ecosystem service model, generating capacity scores based on the suitability of semi-natural habitats to support insect pollinators and travel distances of pollinators.
}
