% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_water_purif.R
\name{capacity_water_purif}
\alias{capacity_water_purif}
\title{Water Purification Capacity Model}
\usage{
capacity_water_purif(
  x,
  studyArea,
  DTM = NULL,
  spr = NULL,
  res = 5,
  threshold = 500,
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attribute HabCode_B.}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{DTM}{A raster of elevation values for the study area. If left NULL (default), will be accessed from the projectLog data. Alternatively a file path can be specified.}

\item{spr}{Basemap attribute name (must be quoted) representing Standard Percentage Runoff values (optional). Default is to leave it out, but if specified it will be included in the scoring.}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{threshold}{Size (m2) below which an isolated patch is not considered able to provide the service. Default is 500 m2.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".
#' @param save Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with capacity scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum capacity for the area).
}
\description{
Runs the air purification ecosystem service model, generating capacity scores based on the ability of the landscape to slow down the flow of water and catch pollutants. Indicators for this model are vegetation roughness and terrain slope (and optionally soil permeability: surface percent runoff data can be obtained from the Hydrology of Soil Types (HOST) dataset.)
}
