% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analysis.R
\name{changeStats}
\alias{changeStats}
\title{Change Statistics}
\usage{
changeStats(pre, post)
}
\arguments{
\item{pre}{A raster or RasterStack object (usually baseline scores).}

\item{post}{A raster or RasterStack object (usually scenario scores).}
}
\value{
A data frame containing the before/after average scores, and the point and percent change, for each service.
}
\description{
Measures the change in average (demand or capacity) score between two sets of service maps. The pre and post maps should already be cropped/masked to the desired extent, and rescaled to the same maximum using the rescale_rasters function.
}
