% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_functions.R
\name{checkcrs}
\alias{checkcrs}
\title{Check Coordinate Reference System}
\usage{
checkcrs(x, target = 27700)
}
\arguments{
\item{x}{Simple features spatial object, or list of sf objects}

\item{target}{EPSG code of desired end projection, default British National Grid, or another sf object whose projection you want to match}
}
\value{
sf object with desired projection
}
\description{
This function checks the coordinate reference system of a spatial object, and transforms to the British National Grid or other desired target if needed.
}
