% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_functions.R
\name{checkgeometry}
\alias{checkgeometry}
\title{Check Geometry}
\usage{
checkgeometry(x, target = "POLYGON")
}
\arguments{
\item{x}{Simple features spatial object, or list of sf objects}

\item{target}{EPSG code of desired end projection, default British National Grid, or another sf object whose projection you want to match}
}
\value{
sf object with desired geometry type.
}
\description{
This function checks the geometry type of a spatial object, and casts to the desired geometry (POLYGON or POINT). Also makes the geometries valid. Used mostly before operations requiring single-part polygon geometry.
}
