% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_functions.R
\name{checkNames}
\alias{checkNames}
\title{Check Attribute Validity}
\usage{
checkNames(folder, layer = NULL, user_names, name = NULL)
}
\arguments{
\item{folder}{Folder where spatial file or files are stored}

\item{layer}{Layer name if several layers present in folder}

\item{user_names}{A named vector of attributes, names being the target EcoservR name and items being the actual name in dataset}

\item{name}{The name of the dataset that is being checked, to customise error message if needed}
}
\value{
A list of spatial files
}
\description{
This function checks that the attribute names supplied by user are indeed present in a dataset without importing it, so that a large object is only read when attributes are valid.
This function is not meant to be called directly by user but called before importing data.
}
