% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_agri}
\alias{classif_agri}
\title{Classify Habitats Using CORINE and/or CROME}
\usage{
classif_agri(x)
}
\arguments{
\item{x}{A basemap sf object, which must contain the attributes HabCode_B and corine or crome.}
}
\value{
The basemap sf object with updated attribute HabCode_B
}
\description{
Revises habitat classification of agricultural land to differentiate pastures from arable land. Uses either CORINE Land Cover, Crop Map of England, or (ideally) both. Not meant to be called directly; rather is used conditionally in classify_habitats().
}
