% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_green}
\alias{classif_green}
\title{Classify Habitats Using OS Greenspace / Open Greenspace}
\usage{
classif_green(x)
}
\arguments{
\item{x}{A basemap sf object, which must contain the attributes HabCode_B and GI}
}
\value{
The basemap sf object with updated attribute HabCode_B
}
\description{
Revises habitat classification by considering green infrastructure information. Not meant to be called directly; rather is used conditionally in classify_habitats().
}
