% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_mastermap}
\alias{classif_mastermap}
\title{Classify Habitats Using OS MasterMap Terms}
\usage{
classif_mastermap(x, params)
}
\arguments{
\item{x}{A basemap sf object, which must contain the attributes Group (Descriptive Group from OS MasterMap), Term (Descriptive Term from OS MasterMap) and Make (from OS MasterMap).}

\item{params}{A list of parameters used as thresholds to classify certain habitats. See EcoservR documentation.}
}
\value{
The basemap sf object with a new attribute HabCode_B
}
\description{
Assigns the most likely habitat type based on information contained by OS MasterMap (using Descriptive Group and Descriptive Term).
}
