% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_nfi}
\alias{classif_nfi}
\title{Classify Habitats Using National Forest Inventory}
\usage{
classif_nfi(x)
}
\arguments{
\item{x}{A basemap sf object, which must contain the attributes HabCode_B and nfi.}
}
\value{
The basemap sf object with updated attribute HabCode_B
}
\description{
Revises habitat classification of woodland. Currently only adds information for felled woodland. Not meant to be called directly; rather is used conditionally in classify_habitats().
}
