% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_phi}
\alias{classif_phi}
\title{Classify Habitats Using OS Priority Habitats Inventory}
\usage{
classif_phi(x)
}
\arguments{
\item{x}{A basemap sf object, which must contain the attributes HabCode_B and phi}
}
\value{
The basemap sf object with updated attribute HabCode_B
}
\description{
Assigns the most likely habitat type based on information contained in priority habitat inventory and first classification step by OS data. Not meant to be called directly; rather is used conditionally in classify_habitats().
}
