% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod09_classify_habitats.R
\name{classify_map}
\alias{classify_map}
\title{Classify Habitats (Phase 1)}
\usage{
classify_map(mm = parent.frame()$mm, projectLog = parent.frame()$projectLog)
}
\arguments{
\item{mm}{The mm object loaded in the environment, can be at various stages of updating.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}
}
\value{
Saves a project_title_MM_classified.RDS file to project folder
}
\description{
This function assigns a habitat type (modified from Phase 1 codes) to each polygon in the basemap according to the information available. Requires at least the basemap preparation function to have run, and optionally (and ideally) also considers CORINE, Priority Habitats, the Crop Map of England, hedgerows, and elevation data.
}
