% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hedgerows.R
\name{clean_hedgerows}
\alias{clean_hedgerows}
\title{Prepare Hedgerows}
\usage{
clean_hedgerows(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{mm}{The mm object loaded in the environment, usually after running the whole basemap process.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a clean_hedges_project-title.RDS file to project folder in intermediary folder. The path is saved in projectLog$clean_hedges and this is the path that is read in the ES models when use_hedges = TRUE.
}
\description{
This function supersedes the add_hedgerows() function that used to embed new geometries within the basemap. Instead, this function imports the source hedgerow data (line data) and cleans it up into a layer that can be used as optional input in the ES models.
}
