% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean Up and Free Memory}
\usage{
cleanUp(scratch = scratch_path)
}
\arguments{
\item{scratch}{The scratch (temporary) folder where outputs are saved}
}
\value{
Nothing; deletes contents of the scratch folder.
}
\description{
This function deletes the contents of the scratch folder and uses garbage collection to return memory to the OS.
}
