% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corine_lookup}
\alias{corine_lookup}
\title{Lookup Table for CORINE Vector Data}
\format{
A named vector with 47 land use types
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc2018}
}
\usage{
corine_lookup
}
\description{
A lookup vector to match the CORINE land use type description (values) to the numeric code (names) used in the CORINE Land Cover 2018 vector dataset.
}
\keyword{datasets}
