% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corine_lookup_raster}
\alias{corine_lookup_raster}
\title{Lookup Table for CORINE Raster Data}
\format{
A data frame with 45 rows and 2 variables:
\describe{
  \item{numcode}{numeric value found in CLC raster data}
  \item{desc}{land use text description to match to numcode}
}
}
\source{
\url{https://land.copernicus.eu/pan-european/corine-land-cover/clc2018}
}
\usage{
corine_lookup_raster
}
\description{
A lookup table to match the CORINE land use type description to the numeric code used in the CORINE Land Cover 2018 raster dataset.
}
\keyword{datasets}
