% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{create_GI}
\alias{create_GI}
\title{Create GI column from MasterMap}
\usage{
create_GI(x, params)
}
\arguments{
\item{x}{A basemap sf object}

\item{params}{A list of parameters used as thresholds to classify certain habitats. See EcoservR documentation.}
}
\value{
The basemap sf object with new attribute GI
}
\description{
If the GI attribute is missing from basemap (e.g. because there was no OS Greenspace / Open Greenspace coverage for a site), this function creates the attribute based on information from OS Mastermap.
}
