% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecol_networks.R
\name{create_network}
\alias{create_network}
\title{Ecological Networks}
\usage{
create_network(
  x = parent.frame()$mm,
  studyArea = parent.frame()$studyArea,
  habitat,
  minsource = 0.1,
  dispersal = NULL,
  threshold = 1,
  use_hedges = FALSE,
  res = 5,
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  export_raster = TRUE,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attribute HabCode_B.}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape.}

\item{habitat}{The habitat for which to compute a network. Must be exactly one of the following: "woodland", "grassland", "heath", "pond", "mire", "wetland"}

\item{minsource}{The minimum patch size considered to be a viable habitat for the generic focal species, in hectares. Default is 0.1 ha (1000 square meters)}

\item{dispersal}{The dispersal distance (in meters) of generic focal species at three scales: core, meso and macro. Do not set the parameter if you wish to use the sensible defaults (see user guide), or specify as a vector of three numeric values, e.g. c(150, 1000, 1500)}

\item{threshold}{Functional size threshold (in hectares) under which a mapped network is considered too small to serve its purpose. Default is 1 ha.}

\item{use_hedges}{Use a separate hedgerow layer? Default FALSE, see clean_hedgerows() for producing a model-ready hedge layer.}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".}

\item{export_raster}{Logical. Save a raster version of the network as well as a polygon layer? Default TRUE.}

\item{save}{Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "networks_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
A polygon and optionally a raster identifying the core, meso, and macro networks with values of 1, 2, and 3 respectively. Note that the scales build on each other (e.g. meso network is the combination of all 1 and 2 values.)
}
\description{
Maps an ecological network to inform connectivity analysis. Six different habitat types can be mapped: woodlands, grasslands, heathlands, lowland wetlands, mires (bogs), and ponds. A generic focal species approach is used and the resulting network is therefore dependent on the habitat requirements (patch size) and dispersal ability of this hypothetical representative species.
}
