% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_access_nature.R
\name{demand_access_nature}
\alias{demand_access_nature}
\title{Access to Nature Demand Model}
\usage{
demand_access_nature(
  x = parent.frame()$mm,
  studyArea = parent.frame()$studyArea,
  res = 5,
  local = 300,
  pop_threshold = 5,
  indicators = TRUE,
  indic_weights = c(1, 0.5, 0.5),
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attributes "HabCode_B", "housePop", "health", "GI", and "GIpublic".}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{local}{Radius (m) for focal statistics at local range, and threshold for distance to greenspace. Default is 300 m.}

\item{pop_threshold}{Population density (people per hectare) below which to consider there is no demand. Default is 5.}

\item{indicators}{Logical; should raw indicators (before transformation into z-scores and rescaling) be saved to the project folder? Default TRUE.}

\item{indic_weights}{A numeric vector of length 3, with weights for distance to greenspaces, population and health, respectively. Default to c(1, 0.5, 0.5) (so that population and health combined contribute the same amount as the ANGSt indicator).}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".}

\item{save}{Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with demand scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum demand for the area).
}
\description{
Runs the access to nature ecosystem service model, generating demand scores based on three indicators: proportion of households meeting the criteria of having accessible greenspaces >2ha within a 300m radius (or specified by user with the "local" argument), population, and health. (Specific indicators can be omitted by setting the appropriate weight to 0.)
}
