% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_climate_reg.R
\name{demand_climate_reg}
\alias{demand_climate_reg}
\title{Climate Regulation Demand Model}
\usage{
demand_climate_reg(
  x,
  studyArea,
  res = 5,
  local = 200,
  urban = NULL,
  urban_size = 10,
  pop_density = 50,
  weights = c(1, 1),
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attribute HabCode_B.}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{local}{Radius (m) for focal statistics at local range (maximum distance for effectiveness). Default is 200 m.}

\item{urban}{Path to a folder containing the Built-up areas layer (ONS Built-up Areas). TO BE BUILT INTO TOOLKIT EVENTUALLY}

\item{urban_size}{Area threshold (km2) for a built-up area to have a heat island effect (default 10 km2).}

\item{pop_density}{Threshold below which population is not considered high enough to have a demand (sum of house pop field in radius of "local"). Default 50.}

\item{weights}{Vector of two: weights to give to societal and regulatory demand, respectively. Default to 1:1.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".
#' @param save Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with capacity scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum capacity for the area).
}
\description{
Runs the climate regulation ecosystem service model, generating capacity scores based on the ability of vegetation and water bodies to cool down the temperature locally.
}
