% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{distance_from_source}
\alias{distance_from_source}
\title{Distance from Source}
\usage{
distance_from_source(x, r, feature, maxdist)
}
\arguments{
\item{x}{the source object(s), as sf polygons}

\item{r}{a template raster of the desired extent and resolution}

\item{feature}{feature name for informative messages}

\item{maxdist}{distance (m) past which the source is considered not to have an effect}
}
\value{
a raster where cell values are distance in meters from sources (which have values of 0)
}
\description{
Generate a raster of distances from source(s), restricted to a maximum search distance.
}
