% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{faster_intersect}
\alias{faster_intersect}
\title{Faster Intersection by Indexing}
\usage{
faster_intersect(x, cliplayer)
}
\arguments{
\item{x}{a simple features object}

\item{cliplayer}{The layer used for clipping, also a simple features object}
}
\value{
A sf object clipped to the desired boundary
}
\description{
This function performs a spatial intersection, but first performs a check to identify only the features that need clipping. Features completely outside the boundary are discarded, and those completely inside are kept as is.
}
