% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_export.R
\name{finalise_basemap}
\alias{finalise_basemap}
\title{Finalise Basemap}
\usage{
finalise_basemap(
  mm = parent.frame()$mm,
  projectLog = parent.frame()$projectLog,
  tiles = FALSE
)
}
\arguments{
\item{mm}{The mm object loaded in the environment. It should have all the desired data added and classified.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{tiles}{Should the basemap be exported as 10x10km tiles? Defaults to FALSE but may be required for large study areas which may take a long time to render in a GIS software.}
}
\value{
Saves a project_title_final.RDS and a project_title_final.gpkg file to project folder.
}
\description{
This function tidies up the final basemap and exports it to the "final" folder. It also outputs summary statistics. (NB: these should be checked, and recalculated if necessary, as MasterMap data may contain some overlapping polygons which will affect the area calculations.)
}
