% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{findLast}
\alias{findLast}
\title{Find Latest Basemap}
\usage{
findLast(dir = NULL, pattern = "MM")
}
\arguments{
\item{dir}{Folder where basemap files are saved}

\item{pattern}{The text pattern to look for, default MM}
}
\value{
The last updated RDS basemap object
}
\description{
This function finds the latest updated file in a directory. Use this to reload your basemap and resumng your project after interrupting a session.
}
