% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{functionalMask}
\alias{functionalMask}
\title{Functional Threshold}
\usage{
functionalMask(x, local = 0, res, proportion = 0.1, threshold)
}
\arguments{
\item{x}{a capacity score raster}

\item{local}{the wide search radius, in meters (defaults to 0 in case it doesn't apply to model)}

\item{res}{the resolution of the raster (m)}

\item{proportion}{the proportion of cells in the search radius which we allow to have a low score (10 units) and still consider 0}

\item{threshold}{the size threshold (m2) for a patch to be considered functional}
}
\value{
the score raster, masked
}
\description{
Create a mask to remove small patches not considered to provide a service. Used in some ecosystem service capacity models. Not meant to be used directly.
}
