% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mastermap.R
\name{getFileExtent}
\alias{getFileExtent}
\title{Get File Extent}
\usage{
getFileExtent(fileList, layer)
}
\arguments{
\item{fileList}{a list of file paths. For shapefiles, should be the dsn path (folder); for other spatial formats, should be the full path including extension.}

\item{layer}{the layer name. For shapefiles, this correspond to the file name and should be a list of same length as fileList. For other spatial formats, layer names can be obtained in a GIS system or by using st_layers() and should be a character string.}
}
\value{
A list of rectangular polygons corresponding to the extent of each file in fileList.
}
\description{
Not intended to be called directly by user. Queries a list of files on disk and extracts only their spatial extent (without reading the file in memory.) Used to associate file names to OS 10km grid tiles for faster Mastermap loading in prepare_basemap().
}
