% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_project_management.R
\name{getProjectLog}
\alias{getProjectLog}
\title{Import Project Log}
\usage{
getProjectLog(wd = NULL)
}
\arguments{
\item{wd}{The working directory - if left out, the function will detect the directory automatically. Alternatively a different project folder can be set, although not recommended.}
}
\value{
The RDS project log
}
\description{
This function finds the project log saved in the working directory
}
