% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mastermap.R
\name{grid_study}
\alias{grid_study}
\title{Grid Study Area}
\usage{
grid_study(x)
}
\arguments{
\item{x}{A sf polygon representing the study area}
}
\value{
A named list of tile boundaries (clipped to study area shape).
}
\description{
Not intended to be called directly by user. Takes a study area and intersects it with the OS 10km x 10km grid to assign a grid reference to each square and enable serial processing of MasterMap.
}
