% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{guessFiletype}
\alias{guessFiletype}
\title{Automatically Detect File Type}
\usage{
guessFiletype(path)
}
\arguments{
\item{path}{Folder where files are stored}
}
\value{
A character string, "gpkg" or "shp", to be fed into loadSpatial
}
\description{
Not intended to be called by user. Used in pre-processing to detect whether spatial files are of a geopackage or shapefile format.
}
