% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hab_lookup}
\alias{hab_lookup}
\title{EcoservR Lookup Table}
\format{
A data frame with XX rows and YY variables.
\describe{
  \item{Ph1code}{EcoservR habitat code to match to the HabCode_B field in the baseline}
  \item{HabNmPLUS}{Expanded description of habitat type}
  \item{HabBroad}{Broad habitat type}
  \item{HabClass}{General habitat class}
  \item{TotCarb}{Average carbon storage capacity for a given habitat (tons per hectare)}
  \item{AirPurScore}{Air purification capacity for a given habitat (0-100)}
  \item{Noise}{Noise regulation capacity for a given habitat (0-100)}
  \item{Rough}{Manning's surface roughness coefficient for a given habitat}
  \item{Naturalness}{Perceived naturalness for a given habitat (0-10)}
  \item{Food}{Food production capacity for a given habitat (0-10), adapted from Alison Smith's 2019 Natural capital in Oxfordshire report}
}
}
\usage{
hab_lookup
}
\description{
A lookup table to add information to the environmental baseline produced by EcoservR, based on the "HabCode_B" field. This is the main table used to assign scores to habitats within the ecosystem service models.
}
\keyword{datasets}
