% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{importDesignatedAreas}
\alias{importDesignatedAreas}
\title{Import Designated Areas}
\usage{
importDesignatedAreas(path, studyArea, dataset)
}
\arguments{
\item{path}{Folder where files are stored}

\item{studyArea}{The study area outline to clip the data to}

\item{dataset}{A character string giving the name of the dataset; used for error messages}
}
\value{
A sf dataframe ready to be used by the model. If no geometries occur in the study area, returns NULL.
}
\description{
Not intended to be called by user. Used in the accessible nature model to bring in external datasets in an automated way.
}
