% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_shiny.R
\name{launchWizard}
\alias{launchWizard}
\title{Launch Project Setup Wizard
This function launches an interactive app to prompt the user to select data inputs and parameters.}
\usage{
launchWizard()
}
\value{
Nothing - but saves an RDS file with the project parameters
}
\description{
Launch Project Setup Wizard
This function launches an interactive app to prompt the user to select data inputs and parameters.
}
