% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cleaning.R
\name{list_to_char}
\alias{list_to_char}
\title{Turn list-column into character column}
\usage{
list_to_char(x)
}
\arguments{
\item{x}{A problematic column}
}
\value{
A vector of character data. Where the original list element was character(0), this is change to NA. Where the original list element had a vector of multiple terms, these are concatenated with a comma.
}
\description{
Turns a list into character format: used in EcoservR to fix some MasterMap attributes that import wrongly from certain format (eg GML).
}
