% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analysis.R
\name{loadRasters}
\alias{loadRasters}
\title{Load Rasters}
\usage{
loadRasters(path, type, pattern = NULL)
}
\arguments{
\item{path}{The path to the folder containing the desired rasters.}

\item{type}{The type of rasters to import (either "capacity" or "demand") - the name of the rasters must contain this string.}

\item{pattern}{An additional optional string to restrict which maps are loaded (for instance if many model runs have been put in the same folder).}
}
\value{
A RasterStack object, with names corresponding to the service
}
\description{
This function imports ecosystem score rasters (capacity or demand as specified by user). Generally this would be for one model run (baseline or intervention scenario). Note that both the type and pattern arguments work with regular expressions so ideally rasters will not have been renamed.
}
