% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling_extract_functions.R
\name{makeTiles}
\alias{makeTiles}
\title{Make Raster Tiles}
\usage{
makeTiles(vectlist, value = NULL, scratch = parent.frame()$scratch_path, name)
}
\arguments{
\item{vectlist}{List of vector tiles, ideally named}

\item{value}{The attribute of interest in the dataset}

\item{scratch}{The path to the temporary scratch folder}

\item{name}{a descriptor passed to temporary files}
}
\value{
A list of raster tiles, named after the mastermap tiles
}
\description{
This function rasterizes a set of vector tiles, usually obtained by prepTiles.
}
