% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plotting.R
\name{panelMaps}
\alias{panelMaps}
\title{Produce Change Maps}
\usage{
panelMaps(x, y, site, sitename, buffersize = 50, nrow = 2)
}
\arguments{
\item{x}{A raster or RasterStack object (usually baseline scores).}

\item{y}{A raster or RasterStack object (usually scenario scores).}

\item{site}{The boundary of the site, as an sf object}

\item{sitename}{The name of the site to appear in the figure caption}

\item{buffersize}{A buffer size (in meters) to apply around the site to give more context to the maps (default 50 m)}

\item{nrow}{Number of rows along which to plot the panels (default 1, so three maps side by side)}
}
\value{
A plot object with scores before and after intervention and the % change per pixel
}
\description{
Creates a panelled figure of before/after and change in ecosystem service scores, overlaid onto a simple black and white baseline map for context.
}
