% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_project_management.R
\name{performance_summary}
\alias{performance_summary}
\title{Performance Summary}
\usage{
performance_summary(
  x = parent.frame()$mm,
  projectLog = parent.frame()$projectLog
)
}
\arguments{
\item{x}{The (final) basemap}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
a HTML document with model run times and information about map size
}
\description{
Outputs an HTML file containing information about the basemap and processing time for each step,
}
