% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_functions.R
\name{permute}
\alias{permute}
\title{Permute Terms}
\usage{
permute(..., sep = ",")
}
\arguments{
\item{...}{Terms to shuffle, in quotation marks}

\item{sep}{The separator that will appear between elements in the resulting string (default is comma without spaces for OS MasterMap terms)}
}
\value{
A vector containing all possible permutations of the terms, each permutation as a character string.
}
\description{
This function returns all possible permutations of given terms.
}
