% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analysis.R
\name{plotifyRasters}
\alias{plotifyRasters}
\title{Plotify Rasters}
\usage{
plotifyRasters(r)
}
\arguments{
\item{r}{A raster or RasterStack object}
}
\value{
A RasterStack object, with names corresponding to the service
}
\description{
Turns rasters into a plottable dataframe for use with ggplot2
}
