% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_import.R
\name{poly_in_boundary}
\alias{poly_in_boundary}
\title{Filtered Import}
\usage{
poly_in_boundary(x, y, layer = NULL)
}
\arguments{
\item{x}{Folder where files are stored}

\item{y}{The study area outline to query the data}

\item{layer}{Name of the layer, if required}
}
\value{
A sf dataframe ready to be used by the model. If no features are intersecting the defined boundary, returns a sf dataframe of length 0.
}
\description{
Helper function to read only features contained in a study area rather than load every feature in the dataset. Called by loadSpatial() for most baseline dataset imports to reduce memory load and speed up computation.
}
