% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod01_prep_mm.R
\name{prepare_basemap}
\alias{prepare_basemap}
\title{Prepare MasterMap}
\usage{
prepare_basemap(projectLog = parent.frame()$projectLog)
}
\arguments{
\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_01.RDS file to project folder, and returns the buffered study area and the mm object to the environment.
}
\description{
This function imports the OS MasterMap from its specified folder. It only reads in polygons that intersect the study area but needs to query all files in the folder to do so. For optimal performance, set your mastermap folder to only contain OS tiles needed for a given project.
}
