% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tiling_extract.R
\name{prepTiles}
\alias{prepTiles}
\title{Prepare Vector Tiles}
\usage{
prepTiles(mm, vect, studyArea = studyAreaBuffer, value)
}
\arguments{
\item{mm}{The mastermap object, as a list of sf objects}

\item{vect}{The vector dataset - sf format - to tile}

\item{studyArea}{The study area boundaries to clip the dataset}

\item{value}{The attribute or attributes of interest in the dataset}
}
\value{
A list of vector tiles, named after the mastermap tiles
}
\description{
This function uses the extent of the mastermap tiles to create vector tiles of the new layer we want to integrate. The tiles are a bit larger than the mastermap tile, ensuring full coverage. If the mastermap tiles are named, so will the vector tiles be, facilitating extraction.
}
