% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_export.R
\name{preview_basemap}
\alias{preview_basemap}
\title{Preview Basemap}
\usage{
preview_basemap(
  mm = parent.frame()$mm,
  studyAreaBuffer = parent.frame()$studyAreaBuffer,
  projectLog = parent.frame()$projectLog,
  sample = NULL
)
}
\arguments{
\item{mm}{The mm object loaded in the environment. It should have all the desired data added and classified.}

\item{studyAreaBuffer}{The buffered study area generated during mod01 or reloaded when resuming a session.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{sample}{A length (in meters) for the side of a sampling square centered on the geographic middle of the map. Useful to export a small chunk of the map for a rapid check rather than exporting the full object. We recommend 5000 m.}
}
\value{
Saves a "project_title_map_preview.gpkg" to project folder.
}
\description{
This function exports the basemap to a spatial format suitable for previewing in a GIS software. It is useful to frequently check on the outputs during the basemapping process to make sure that the coverage is as expected and that the data extracted look reasonable. NB: This is NOT the way to export the final map: use finalise_basemap for that.
}
