% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{raster_centroid}
\alias{raster_centroid}
\title{Find center of a raster}
\usage{
raster_centroid(r)
}
\arguments{
\item{r}{The raster object}
}
\value{
A sf point
}
\description{
Calculates the middle point of a raster from its extent. Assumes the projection is British National Grid (EPSG 27700). Not to be called directly: used in the add_DTM() workflow to assign a grid reference to tiles.
}
