% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_checking.R
\name{removeDuplicPoly}
\alias{removeDuplicPoly}
\title{Remove Duplicated Polygons}
\usage{
removeDuplicPoly(tiles, ID)
}
\arguments{
\item{tiles}{List of simple features spatial objects, usually OS MasterMap tiles}

\item{ID}{Attribute containing a polygon's unique identifier}
}
\value{
The same list with duplicates removed
}
\description{
This function loops through a list of tiles and removes polygons that are duplicated across tiles. If only one tile is used, the distinct function.
}
