% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename geometry column}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{The sf object whose geometry you want to change}

\item{name}{The name you want the geometry column renamed to}
}
\value{
the sf object g with the geometry column name changed
}
\description{
The geometry column of sf object can have inconsistent names (e.g. geom, geometry, etc). This function renames the geometry attribute to a given name.
}
