% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_spatial.R
\name{rescale_indicator}
\alias{rescale_indicator}
\title{Rescale Indicator}
\usage{
rescale_indicator(r)
}
\arguments{
\item{r}{a RasterLayer with values to transform}
}
\value{
the raster with standardised and rescaled values
}
\description{
Converts a score raster (usually a demand indicator) to z-scores (mean-centered and variance-scaled) and rescales them 0-1.
}
