% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analysis.R
\name{rescale_rasters}
\alias{rescale_rasters}
\title{Rescale Rasters}
\usage{
rescale_rasters(x, y)
}
\arguments{
\item{x}{A raster or RasterStack object (usually baseline scores).}

\item{y}{A raster or RasterStack object (usually scenario scores).}
}
\value{
A list of two Raster or RasterStack objects, rescaled 0-100 using a common maximum (per service).
}
\description{
Rasters from different model runs (e.g. pre and post intervention scenario) may have different score ranges. It is important to rescale them using a common maximum so we can measure the change.
}
