% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_functions.R
\name{resumeProject}
\alias{resumeProject}
\title{Resume Existing Project}
\usage{
resumeProject(folder = NULL, trustlog = FALSE)
}
\arguments{
\item{folder}{Project folder (by default the working directory)}

\item{trustlog}{Should the project log decide which MM object to load? If TRUE, looks for the file path recorded in the last_success slot even if a more recent version exists. If FALSE, prefers the MM file with the latest timestamp.}
}
\value{
The projectLog object, the mm object, and the studyAreaBuffer object. Also prints an indication of what datasets are already incorporated and which remain.
}
\description{
This function checks a folder for the latest project log, and once it finds it loads in the environment the latest mm object and the buffered study area.
}
