% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_classification.R
\name{classif_area}
\alias{classif_area}
\title{Classify Habitats Using Size Criteria}
\usage{
classif_area(x, params)
}
\arguments{
\item{x}{A basemap sf object}

\item{params}{A list of parameters used as thresholds to classify certain habitats. See EcoservR documentation.}
}
\value{
The basemap sf object with updated attribute HabCode_B
}
\description{
Revises habitat classification of agricultural land using the parameters set with the project. Not meant to be called directly; rather is used conditionally in classify_habitats().
}
