% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean Up and Free Memory}
\usage{
cleanUp(scratch = scratch_path)
}
\arguments{
\item{folder}{Folder where spatial file or files are stored}

\item{layer}{Layer name if several layers present in folder}

\item{filetype}{File format, either "gpkg" for geopackage or "shp" for shapefile}
}
\value{
A list of spatial files
}
\description{
This function deletes the contents of the scratch folder and uses garbage collection to return memory to the OS.
}
