% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_climate_reg.R
\name{demand_climate_reg}
\alias{demand_climate_reg}
\title{Climate Regulation Demand Model}
\usage{
demand_climate_reg(
  x = parent.frame()$mm,
  studyArea = parent.frame()$studyArea,
  res = 5,
  local = 200,
  urban = NULL,
  urban_size = 10,
  pop_density = 50,
  indicators = TRUE,
  indic_weights = c(1, 1),
  projectLog = parent.frame()$projectLog,
  runtitle = parent.frame()$runtitle,
  save = NULL
)
}
\arguments{
\item{x}{A basemap, in a list of sf tiles or as one sf object. Must have attributes "HabCode_B", "housePop", "riskgroup".}

\item{studyArea}{The boundaries of the site, as one sf object. The final raster will be masked to this shape. For best results this shape should be smaller than the basemap (which should be buffered by typically 300 m - 1km to avoid edge effects).}

\item{res}{Desired resolution of the raster. Default is 5 m. Range recommended is 5-10m.}

\item{local}{Radius (m) for focal statistics at local range. Default is 200 m.}

\item{urban}{Folder path to the Built-up Areas dataset (https://data.gov.uk/dataset/ad0381f2-322b-4699-94ec-32839ec02d33/built-up-areas-december-2011-boundaries-v2).}

\item{urban_size}{Area threshold (km2) for a builtup area to have a heat island effect (default 10 km2).}

\item{pop_density}{Population threshold (number of residents within "local" distance) above which we consider there is a demand (default 50).}

\item{indicators}{Logical; should raw indicators (before transformation into z-scores and rescaling) be saved to the project folder? Default TRUE.}

\item{indic_weights}{A numeric vector of length 4, with weights for distance to roads, manmade surfaces, population and health, respectively. Default to equal weights of 1 (all indicators contributing equally to final sum).}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{runtitle}{A customised title you can give a specific model run, which will be appended to your project title in the outputs. If comparing a basemap to an intervention map, we recommend using "pre" and "post", or a short description of the interventions, e.g. "baseline" vs "tree planting".}

\item{save}{Path to folder where outputs will be saved. By default a folder will be created using your chosen run title, prefixed by "services_". Do not use this argument unless you need to save the outputs somewhere else.}
}
\value{
Two rasters with demand scores: one with raw scores (arbitrary units), and one rescaled 0-100 (where 100 is maximum demand for the area).
}
\description{
Runs the climate regulation ecosystem service model, generating demand scores based on two indicators: population (weighted by age risk group)  and proportion of manmade surfaces. (Specific indicators can be omitted by setting the appropriate weight to 0.)
}
