% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_export.R
\name{finalise_basemap}
\alias{finalise_basemap}
\title{Finalise Basemap}
\usage{
finalise_basemap(
  mm = parent.frame()$mm,
  projectLog = parent.frame()$projectLog,
  export = TRUE,
  format = "gpkg",
  tiles = FALSE
)
}
\arguments{
\item{mm}{The mm object loaded in the environment. It should have all the desired data added and classified.}

\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}

\item{export}{Should a clipped, recombined version of the map be saved as a GIS file? Default TRUE}

\item{format}{The desired output format, only used if export = TRUE. Currently supported are "gpkg" (geopackage, the default), "ESRI Shapefile" and "MapInfo File".}

\item{tiles}{Should the basemap be exported as 10x10km tiles? Defaults to FALSE but may be required for large study areas which may take a long time to render in a GIS software. Only used when export = TRUE.}
}
\value{
Saves a project_title_final.RDS and optionally a project_title_final GIS file to project folder.
}
\description{
This function tidies up the final basemap and exports it to the "final" folder. It also outputs summary statistics. (NB: these should be checked, and recalculated if necessary, as MasterMap data may contain some overlapping polygons which will affect the area calculations.)
}
