% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_import.R
\name{loadSpatial}
\alias{loadSpatial}
\title{Load Spatial Data}
\usage{
loadSpatial(folder, layer = NULL, filetype)
}
\arguments{
\item{folder}{Folder where spatial file or files are stored}

\item{layer}{Layer name if several layers present in folder}

\item{filetype}{File format. Accepted file extensions are "gpkg", "shp", "json", "gz", "gml", "tab".}
}
\value{
A list of spatial files
}
\description{
This function looks into a folder for a specified file type and loads all corresponding files into a list.
}
