% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod01_prep_mm.R
\name{prepare_basemap}
\alias{prepare_basemap}
\title{Prepare MasterMap}
\usage{
prepare_basemap(projectLog = parent.frame()$projectLog)
}
\arguments{
\item{projectLog}{The RDS project log file generated by the wizard app and containing all file paths to data inputs and model parameters}
}
\value{
Saves a project_title_MM_01.RDS file to project folder
}
\description{
This function imports the OS MasterMap from its specified folder.
}
