% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{updateProjectLog}
\alias{updateProjectLog}
\title{Update Project Log}
\usage{
updateProjectLog(projlog)
}
\arguments{
\item{projlog}{The updated project log to be saved}
}
\value{

}
\description{
This function finds the project log saved in the working directory and overwrites it with the new log
}
